Update expr 1 to split exper and analysis loading
Now that experiments and analyses are separated to an extent it
doesnt' make sense to load both at once. All the experiment data
loading scripts need to be updated so that loading data related
to the experiment as a whole is separate from loading analysis
results.

+++ utils/experiment.py
@@ -1,13 +1,13 @@
 import json
 import os.path
 from datetime import datetime, timezone
-from typing import IO, Any, TypeVar
+from typing import IO, Any, Optional, TypeVar
 
-from cegs_portal.search.conftest import experiment
 from cegs_portal.search.models import (
     AccessionIdLog,
     AccessionIds,
     AccessionType,
+    Analysis,
     Biosample,
     CellLine,
     Experiment,
@@ -33,9 +33,7 @@ class ExperimentFileMetadata:
         self.significance_measure = file_metadata["significance_measure"]
         self.p_val_threshold = file_metadata.get("p_val_threshold", 0.05)
 
-    def db_save(self, experiment: Experiment):
-        self.file_metadata.db_save(experiment)
-
+    def db_save(self, experiment: Experiment, analysis: Analysis = None):
         data_file_info = ExperimentDataFileInfo(
             ref_genome=self.ref_genome,
             ref_genome_patch=self.ref_genome_patch,
@@ -43,6 +41,7 @@ class ExperimentFileMetadata:
             p_value_threshold=self.p_val_threshold,
         )
         data_file_info.save()
+        self.file_metadata.db_save(experiment, analysis, data_file_info)
 
 
 T = TypeVar("T", bound="ExperimentMetadata")
@@ -62,9 +61,9 @@ class ExperimentBiosample:
         self.experiment = experiment
 
     def db_save(self):
-        with AccessionIds(message=f"{experiment.accession_id}: {experiment.name}"[:200]) as accession_ids:
-            cell_line = CellLine.objects.filter(name=self.cell_line).first()
-            if cell_line is None:
+        cell_line = CellLine.objects.filter(name=self.cell_line).first()
+        if cell_line is None:
+            with AccessionIds(message=f"{self.experiment.accession_id}: {self.experiment.name}"[:200]) as accession_ids:
                 tissue_type, tt_created = TissueType.objects.get_or_create(name=self.tissue_type)
                 if tt_created:
                     tissue_type.accession_id = accession_ids.incr(AccessionType.TT)
@@ -76,7 +75,11 @@ class ExperimentBiosample:
                     tissue_type_name=self.tissue_type,
                 )
                 cell_line.save()
-
+        bios = Biosample.objects.filter(
+            cell_line=cell_line,
+            cell_line_name=cell_line.name,
+        ).first()
+        if bios is None:
             bios = Biosample(
                 cell_line=cell_line,
                 cell_line_name=cell_line.name,
@@ -85,15 +88,71 @@ class ExperimentBiosample:
             bios.save()
 
 
+class AnalysisMetadata:
+    accession_id: Optional[str] = None
+    experiment_accession_id: str
+    filename: str
+    description: str
+    name: str
+    results: list[ExperimentFileMetadata] = []
+
+    def __init__(self, analysis_dict: dict[str, Any], analysis_filename: str):
+        self.description = analysis_dict["description"]
+        self.experiment_accession_id = analysis_dict["experiment"]
+        self.filename = analysis_filename
+        self.name = analysis_dict["name"]
+
+        for result in analysis_dict["results"]:
+            self.results.append(ExperimentFileMetadata(result))
+
+    def db_save(self):
+        experiment = Experiment.objects.get(accession_id=self.experiment_accession_id)
+        with AccessionIds(message=f"Analysis of {experiment.accession_id}: {experiment.name}"[:200]) as accession_ids:
+            analysis = Analysis(description=self.description, experiment=experiment, name=self.name)
+            analysis.accession_id = accession_ids.incr(AccessionType.ANALYSIS)
+            analysis.save()
+            self.accession_id = analysis.accession_id
+        for result in self.results:
+            result.db_save(experiment, analysis)
+
+        return analysis
+
+    def db_del(self, analysis=None):
+        if self.accession_id is None and analysis is None:
+            print("Analysis has not been saved: There is no accession ID")
+            return
+
+        if self.accession_id is not None:
+            analysis = Analysis.objects.get(accession_id=self.accession_id)
+
+        for file in analysis.files.all():
+            file.data_file_info.delete()
+        analysis.delete()
+        self.accession_id = None
+
+    def metadata(self):
+        base_path = os.path.dirname(self.filename)
+        for result in self.results:
+            delimiter = get_delimiter(result.file_metadata.filename)
+            data_file = open(os.path.join(base_path, result.file_metadata.filename), "r", newline="")
+            yield data_file, result, delimiter
+            data_file.close()
+
+    @classmethod
+    def json_load(cls, file: IO):
+        analysis_data = json.load(file)
+        metadata = AnalysisMetadata(analysis_data, file.name)
+        return metadata
+
+
 class ExperimentMetadata:
-    file_metadata: list[ExperimentFileMetadata]
     description: str
     experiment_type: str
     name: str
     filename: str
     accession_id: str
     biosamples: list[ExperimentBiosample]
-    other_file_metadata: list[FileMetadata]
+    file_metadata: list[FileMetadata]
 
     def __init__(self, experiment_dict: dict[str, Any], experiment_filename: str):
         self.description = experiment_dict.get("description", None)
@@ -102,12 +161,10 @@ class ExperimentMetadata:
         self.accession_id = experiment_dict["accession_id"]
         self.filename = experiment_filename
         self.file_metadata = []
-        self.other_file_metadata = []
+        self.file_metadata = []
         self.biosamples = [ExperimentBiosample(sample, self) for sample in experiment_dict["biosamples"]]
-        for data in experiment_dict["data"]:
-            self.file_metadata.append(ExperimentFileMetadata(data))
-        for file in experiment_dict.get("other_files", []):
-            self.other_file_metadata.append(FileMetadata(file))
+        for file in experiment_dict.get("files", []):
+            self.file_metadata.append(FileMetadata(file))
 
     def db_save(self):
         experiment = Experiment(
@@ -117,9 +174,7 @@ class ExperimentMetadata:
             experiment_type=self.experiment_type,
         )
         experiment.save()
-        for metadata in self.file_metadata:
-            metadata.db_save(experiment)
-        for file in self.other_file_metadata:
+        for file in self.file_metadata:
             file.db_save(experiment)
         accession_log = AccessionIdLog(
             created_at=datetime.now(timezone.utc),
@@ -135,9 +190,11 @@ class ExperimentMetadata:
     def db_del(self):
         experiment = Experiment.objects.get(accession_id=self.accession_id)
         for file in experiment.files.all():
-            for data in file.data_file_info.all():
-                data.delete()
+            if file.data_file_info is not None:
+                for data in file.data_file_info.all():
+                    data.delete()
             file.delete()
+
         experiment.delete()
 
     def metadata(self):
